Added Google Places API; EditLocationVC
I copied over the EditLocationVC from the OWTPhotoChecklist app to use for adding locations to and editing locations in the different walking tours.  The map view on the EditLocationVC will center itself depending on which city is selected on the previous view controller.

I also added the Google Places API to utilize the GWSAutocompleteViewController to search for locations/addresses to add.  It works a lot better than what I had in the OWTPhotoChecklist app.  Once a location is selected from the autocomplete form, the map re-centers on the selected location and drops a marker, and the appropriate text fields are filled in.

Still left to do: add a mechanism in the MapVC to allow editing of a location already in the tour, and passing that location's information to the EditLocationVC via segue.